fix: export async interface for stream types
