+++ crates/rpc/src/stream_types.rs
@@ -176,8 +176,16 @@ impl RequestStream {
     }
 }
 
-#[cfg_attr(feature = "uniffi", uniffi::export)]
+#[cfg(feature = "uniffi")]
+#[uniffi::export]
 impl RequestStream {
+    /// Pull the next message from the stream (blocking version)
+    ///
+    /// Returns a StreamMessage indicating the result
+    pub fn next(&self) -> StreamMessage {
+        crate::get_runtime().block_on(self.next_async())
+    }
+
     /// Pull the next message from the stream (async version)
     ///
     /// Returns a StreamMessage indicating the result
@@ -229,8 +237,36 @@ impl ResponseSink {
     }
 }
 
-#[cfg_attr(feature = "uniffi", uniffi::export)]
+#[cfg(feature = "uniffi")]
+#[uniffi::export]
 impl ResponseSink {
+    /// Send a message to the response stream (blocking version)
+    ///
+    /// Returns an error if the stream has been closed or if sending fails.
+    pub fn send(&self, data: Vec<u8>) -> Result<(), RpcError> {
+        crate::get_runtime().block_on(self.send_async(data))
+    }
+
+    /// Send an error to the response stream and close it (blocking version)
+    ///
+    /// This terminates the stream with an error status.
+    pub fn send_error(&self, error: RpcError) -> Result<(), RpcError> {
+        crate::get_runtime().block_on(self.send_error_async(error))
+    }
+
+    /// Close the response stream (blocking version)
+    ///
+    /// Signals that no more messages will be sent.
+    /// The stream will end gracefully.
+    pub fn close(&self) -> Result<(), RpcError> {
+        crate::get_runtime().block_on(self.close_async())
+    }
+
+    /// Check if the sink has been closed (blocking version)
+    pub fn is_closed(&self) -> bool {
+        crate::get_runtime().block_on(self.is_closed_async())
+    }
+
     /// Send a message to the response stream (async version)
     ///
     /// Returns an error if the stream has been closed or if sending fails.
@@ -297,8 +333,16 @@ impl ResponseStreamReader {
     }
 }
 
-#[cfg_attr(feature = "uniffi", uniffi::export)]
+#[cfg(feature = "uniffi")]
+#[uniffi::export]
 impl ResponseStreamReader {
+    /// Pull the next message from the response stream (blocking version)
+    ///
+    /// Returns a StreamMessage indicating the result
+    pub fn next(&self) -> StreamMessage {
+        crate::get_runtime().block_on(self.next_async())
+    }
+
     /// Pull the next message from the response stream (async version)
     ///
     /// Returns a StreamMessage indicating the result
@@ -363,8 +407,19 @@ impl RequestStreamWriter {
     }
 }
 
-#[cfg_attr(feature = "uniffi", uniffi::export)]
+#[cfg(feature = "uniffi")]
+#[uniffi::export]
 impl RequestStreamWriter {
+    /// Send a request message to the stream (blocking version)
+    pub fn send(&self, data: Vec<u8>) -> Result<(), RpcError> {
+        crate::get_runtime().block_on(self.send_async(data))
+    }
+
+    /// Finalize the stream and get the response (blocking version)
+    pub fn finalize_stream(&self) -> Result<Vec<u8>, RpcError> {
+        crate::get_runtime().block_on(self.finalize_stream_async())
+    }
+
     /// Send a request message to the stream (async version)
     pub async fn send_async(&self, data: Vec<u8>) -> Result<(), RpcError> {
         let sender = self.sender.lock().await;
@@ -402,7 +457,8 @@ impl RequestStreamWriter {
     }
 }
 
-#[cfg_attr(feature = "uniffi", uniffi::export)]
+#[cfg(feature = "uniffi")]
+#[uniffi::export]
 impl RequestStreamWriter {
     /// Finalize the stream and get the response (async version)
     ///
@@ -465,8 +521,24 @@ impl BidiStreamHandler {
     }
 }
 
-#[cfg_attr(feature = "uniffi", uniffi::export)]
+#[cfg(feature = "uniffi")]
+#[uniffi::export]
 impl BidiStreamHandler {
+    /// Send a request message to the stream (blocking version)
+    pub fn send(&self, data: Vec<u8>) -> Result<(), RpcError> {
+        crate::get_runtime().block_on(self.send_async(data))
+    }
+
+    /// Close the request stream (no more messages will be sent)
+    pub fn close_send(&self) -> Result<(), RpcError> {
+        crate::get_runtime().block_on(self.close_send_async())
+    }
+
+    /// Receive the next response message (blocking version)
+    pub fn recv(&self) -> StreamMessage {
+        crate::get_runtime().block_on(self.recv_async())
+    }
+
     /// Send a request message to the stream (async version)
     pub async fn send_async(&self, data: Vec<u8>) -> Result<(), RpcError> {
         let sender = self.sender.lock().await;
@@ -575,8 +647,14 @@ impl MulticastResponseReader {
     }
 }
 
-#[cfg_attr(feature = "uniffi", uniffi::export)]
+#[cfg(feature = "uniffi")]
+#[uniffi::export]
 impl MulticastResponseReader {
+    /// Pull the next item from the multicast response stream (blocking).
+    pub fn next(&self) -> MulticastStreamMessage {
+        crate::get_runtime().block_on(self.next_async())
+    }
+
     /// Pull the next item from the multicast response stream (async).
     pub async fn next_async(&self) -> MulticastStreamMessage {
         let mut rx = self.inner.lock().await;
@@ -645,8 +723,25 @@ impl MulticastBidiStreamHandler {
     }
 }
 
-#[cfg_attr(feature = "uniffi", uniffi::export)]
+#[cfg(feature = "uniffi")]
+#[uniffi::export]
 impl MulticastBidiStreamHandler {
+    /// Send a request message to the stream (blocking).
+    pub fn send(&self, data: Vec<u8>) -> Result<(), RpcError> {
+        crate::get_runtime().block_on(self.send_async(data))
+    }
+
+    /// Close the request stream — signals that no more messages will be sent
+    /// (blocking).
+    pub fn close_send(&self) -> Result<(), RpcError> {
+        crate::get_runtime().block_on(self.close_send_async())
+    }
+
+    /// Receive the next response item (blocking).
+    pub fn recv(&self) -> MulticastStreamMessage {
+        crate::get_runtime().block_on(self.recv_async())
+    }
+
     /// Send a request message to the stream (async).
     pub async fn send_async(&self, data: Vec<u8>) -> Result<(), RpcError> {
         let sender = self.sender.lock().await;
